+++ lib/strava/models/activity_total.rb
@@ -1,0 +1,21 @@
+module Strava
+  module Models
+    class ActivityTotal < Model
+      include Mixins::Distance
+      include Mixins::Time
+      include Mixins::Elevation
+
+      property 'count'
+      property 'achievement_count'
+      property 'elevation_gain'
+
+      def total_elevation_gain
+        elevation_gain
+      end
+
+      def units
+        :metric
+      end
+    end
+  end
+end